Created README.md
Added a basic outline of all points listed in the assignment. Added my name to the roles relevant to me, allowing the remaining people to add their own. Add more info as you like. I am also unsure if this is all I need to do to count as my contribution or if my name is needed on evvvvvvvery line of it... "Make sure all contributions you made to the README file have your name listed. If your name is not listed in the README file, you will not receive credit for the assignment."-Module 2 assignment rubric.